+++ zcash_client_backend/src/data_api/testing/transparent.rs
@@ -5,6 +5,7 @@ use assert_matches::assert_matches;
 use ::transparent::{
     address::TransparentAddress,
     bundle::{OutPoint, TxOut},
+    keys::TransparentKeyScope,
 };
 use sapling::zip32::ExtendedSpendingKey;
 use zcash_keys::{
@@ -120,8 +121,13 @@ where
     let txout = TxOut::new(value, taddr.script().into());
 
     // Pretend the output's transaction was mined at `height_1`.
-    let utxo = WalletTransparentOutput::from_parts(outpoint.clone(), txout.clone(), Some(height_1))
-        .unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        outpoint.clone(),
+        txout.clone(),
+        Some(height_1),
+        Some(TransparentKeyScope::EXTERNAL),
+    )
+    .unwrap();
     let res0 = st.wallet_mut().put_received_transparent_utxo(&utxo);
     assert_matches!(res0, Ok(_));
 
@@ -146,7 +152,13 @@ where
     // the same `UtxoId`.
     let height_2 = birthday + 34567;
     st.wallet_mut().update_chain_tip(height_2).unwrap();
-    let utxo2 = WalletTransparentOutput::from_parts(outpoint, txout, Some(height_2)).unwrap();
+    let utxo2 = WalletTransparentOutput::from_parts(
+        outpoint,
+        txout,
+        Some(height_2),
+        Some(TransparentKeyScope::EXTERNAL),
+    )
+    .unwrap();
     let res1 = st.wallet_mut().put_received_transparent_utxo(&utxo2);
     assert_matches!(res1, Ok(id) if id == res0.unwrap());
 
@@ -226,7 +238,13 @@ where
 
     // Pretend the output was received in the chain tip.
     let height = st.wallet().chain_height().unwrap().unwrap();
-    let utxo = WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height)).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        OutPoint::fake(),
+        txout,
+        Some(height),
+        Some(TransparentKeyScope::EXTERNAL),
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -367,7 +385,13 @@ where
 
     // Pretend the output was received in the chain tip.
     let height = st.wallet().chain_height().unwrap().unwrap();
-    let utxo = WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height)).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(
+        OutPoint::fake(),
+        txout,
+        Some(height),
+        Some(TransparentKeyScope::EXTERNAL),
+    )
+    .unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -733,7 +757,7 @@ where
     let value = Zatoshis::const_from_u64(50_000);
     let outpoint = OutPoint::fake();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height)).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height), None).unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -804,7 +828,8 @@ where
     let value = Zatoshis::from_u64(100000).unwrap();
     let height = st.wallet().chain_height().unwrap().unwrap();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height)).unwrap();
+    let utxo =
+        WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height), None).unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -1058,7 +1083,7 @@ where
     let value = Zatoshis::const_from_u64(50_000);
     let outpoint = OutPoint::fake();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height)).unwrap();
+    let utxo = WalletTransparentOutput::from_parts(outpoint, txout, Some(height), None).unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();
@@ -1126,7 +1151,8 @@ where
     let value = Zatoshis::from_u64(100000).unwrap();
     let height = st.wallet().chain_height().unwrap().unwrap();
     let txout = TxOut::new(value, taddr.script().into());
-    let utxo = WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height)).unwrap();
+    let utxo =
+        WalletTransparentOutput::from_parts(OutPoint::fake(), txout, Some(height), None).unwrap();
     st.wallet_mut()
         .put_received_transparent_utxo(&utxo)
         .unwrap();