+++ zcash_client_memory/src/input_source.rs
@@ -1,7 +1,5 @@
 use std::num::NonZeroU32;
 
-#[cfg(feature = "transparent-inputs")]
-use zcash_client_backend::data_api::WalletUtxo;
 use zcash_client_backend::{
     data_api::{
         AccountMeta, InputSource, NoteFilter, PoolMeta, ReceivedNotes, TargetValue, WalletRead,
@@ -22,7 +20,7 @@ use zcash_protocol::ShieldedProtocol::Orchard;
 #[cfg(feature = "transparent-inputs")]
 use {
     ::transparent::{address::TransparentAddress, bundle::OutPoint},
-    zcash_client_backend::data_api::TransactionStatus,
+    zcash_client_backend::{data_api::TransactionStatus, wallet::WalletTransparentOutput},
     zcash_protocol::consensus::BlockHeight,
 };
 
@@ -154,7 +152,7 @@ impl<P: consensus::Parameters> InputSource for MemoryWalletDb<P> {
         address: &TransparentAddress,
         target_height: TargetHeight,
         confirmations_policy: ConfirmationsPolicy,
-    ) -> Result<Vec<WalletUtxo>, Self::Error> {
+    ) -> Result<Vec<WalletTransparentOutput>, Self::Error> {
         // TODO: take into consideration coinbase maturity
         // See <https://github.com/zcash/librustzcash/issues/821>
         let txos = self
@@ -168,14 +166,6 @@ impl<P: consensus::Parameters> InputSource for MemoryWalletDb<P> {
             })
             .filter_map(|(outpoint, txo, tx)| {
                 txo.to_wallet_transparent_output(outpoint, tx.and_then(|tx| tx.mined_height()))
-                    .map(|out| {
-                        WalletUtxo::new(
-                            out,
-                            // FIXME: this needs to be updated to identify the transparent key
-                            // scope for derived addresses in the wallet.
-                            None,
-                        )
-                    })
             })
             .collect();
         Ok(txos)
@@ -186,7 +176,7 @@ impl<P: consensus::Parameters> InputSource for MemoryWalletDb<P> {
         &self,
         outpoint: &OutPoint,
         _target_height: TargetHeight,
-    ) -> Result<Option<WalletUtxo>, Self::Error> {
+    ) -> Result<Option<WalletTransparentOutput>, Self::Error> {
         // FIXME: make use of `target_height` to check spendability.
         Ok(self
             .transparent_received_outputs
@@ -194,14 +184,6 @@ impl<P: consensus::Parameters> InputSource for MemoryWalletDb<P> {
             .map(|txo| (txo, self.tx_table.get(&txo.transaction_id)))
             .and_then(|(txo, tx)| {
                 txo.to_wallet_transparent_output(outpoint, tx.and_then(|tx| tx.mined_height()))
-                    .map(|out| {
-                        WalletUtxo::new(
-                            out,
-                            // FIXME: this needs to be updated to identify the transparent key
-                            // scope for derived addresses in the wallet.
-                            None,
-                        )
-                    })
             }))
     }